Ajustar configuraciones para agregar nuevo empleado y actualizar empl…

+++ client/src/services/deleteEmployee.js
@@ -1,4 +1,7 @@
 export const deleteEmployee = async (employeeId) => {
+  if (!employeeId || isNaN(employeeId)) {
+    throw new Error("El id del empleado es incorrecto");
+  }
   try {
     const options = {
       method: "DELETE",
@@ -17,6 +20,7 @@ export const deleteEmployee = async (employeeId) => {
       const errorData = await response.json();
       throw new Error(errorData.message || "Error al eliminar el empleado");
     }
+    return true;
   } catch (error) {
     console.error("Error al eliminar empleado:", error);
     throw error;
